pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 70.1", "setuptools_scm[toml] ~= 8.1"]

[dependency-groups]
dev = [
  "deptry>=0.23",
  "mkdocs-material>=9.6",
  "mkdocs>=1.6",
  "mkdocstrings[python]>=0.30",
  "mock",
  "mypy>=1.17",
  "pre-commit>=3.8",
  "pytest-cov>=4.1",
  "pytest>=7.4",
  "restview",
  "ruff>=0.12",
  "sphinx",
  "sphinx_rtd_theme",
  "tox-uv>=1.28",
  "vcrpy"
]

[project]
authors = [
  {email = "biocommons-dev@googlegroups.com", name = "biocommons contributors"}
]
classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "lxml",
  "pytz",
  "requests"
]
description = "Python interface to NCBI's eutilities API"
dynamic = ["version"]
license = "Apache-2.0"
name = "biocommons.eutils"
readme = "README.rst"
requires-python = ">=3.11"

[project.urls]
Documentation = "https://eutils.readthedocs.org"
Homepage = "https://github.com/biocommons/eutils"
Issues = "https://github.com/biocommons/eutils/issues"
Repository = "https://github.com/biocommons/eutils"

[tool.coverage.report]
exclude_lines = [
  "def __repr__",
  "if __name__ == .__main__.:",
  "if self.debug",
  "pragma: no cover",
  "raise AssertionError",
  "raise NotImplementedError"
]
show_missing = true
skip_empty = true

[tool.coverage.run]
branch = true
omit = ["*/test/*", "*/tests/*", "*_test.py"]
source = ["src"]

[tool.deptry]

[tool.deptry.package_module_name_map]
# map package name to import name
# Making this explicit suppresses deptry notices
mkdocs = "mkdocs"
mkdocs-material = "mkdocs_material"
mkdocstrings = "mkdocstrings"
mypy = "mypy"
pre-commit = "pre_commit"
pytest = "pytest"
pytest-cov = "pytest_cov"
pyyaml = "yaml"
ruff = "ruff"
tox-uv = "tox_uv"
ty = "ty"

[tool.pytest.ini_options]
addopts = "-s -v -x --strict-markers -m 'not extra' --doctest-modules --cov=src"
doctest_optionflags = [
  "ALLOW_BYTES",
  "ALLOW_UNICODE",
  "ELLIPSIS",
  "IGNORE_EXCEPTION_DETAIL",
  "NORMALIZE_WHITESPACE"
]
markers = [
  "network: tests that require network connectivity",
  "slow: slow tests that should be run infrequently"
]
testpaths = ["tests"]

[tool.ruff]
exclude = [
  "docs"
]
fix = true
line-length = 100
src = ["src", "tests"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
fixable = [
  "B",
  "C4",
  "D",
  "EM",
  "F401",
  "F541",
  "I",
  "PERF",
  "PIE",
  "PT",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "UP"
]
ignore = [
  "E111",
  "E114",
  "E117",
  "E501",
  "E731",
  "PLR0913",
  "S321",
  "W191"
]
select = [
  "A", # https://docs.astral.sh/ruff/rules/#flake8-builtins-a
  "ARG", # https://docs.astral.sh/ruff/rules/#flake8-unused-arguments-arg
  "B", # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
  "C4", # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
  "DTZ", # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
  "E",
  "EM", # https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
  "F", # https://docs.astral.sh/ruff/rules/#pyflakes-f
  "G", # https://docs.astral.sh/ruff/rules/#flake8-logging-format-g
  "I", # https://docs.astral.sh/ruff/rules/#isort-i
  "LOG", # https://docs.astral.sh/ruff/rules/#flake8-logging-log
  "N", # https://docs.astral.sh/ruff/rules/#pep8-naming-n
  "PERF", # https://docs.astral.sh/ruff/rules/#perflint-perf
  "PIE", # https://docs.astral.sh/ruff/rules/#flake8-pie-pie
  "PL", # https://docs.astral.sh/ruff/rules/#pylint-pl
  "PT", # https://docs.astral.sh/ruff/rules/#flake8-pytest-style-pt
  "PTH", # https://docs.astral.sh/ruff/rules/#flake8-use-pathlib-pth
  "RET", # https://docs.astral.sh/ruff/rules/#flake8-return-ret
  "RSE", # https://docs.astral.sh/ruff/rules/#flake8-raise-rse
  "RUF", # https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
  "S", # https://docs.astral.sh/ruff/rules/#flake8-bandit-s
  "SIM", # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
  "TRY", # https://docs.astral.sh/ruff/rules/#tryceratops-try
  "UP", # https://docs.astral.sh/ruff/rules/#pyupgrade-up
  "W", # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
  "YTT" # https://docs.astral.sh/ruff/rules/#flake8-2020-ytt
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "S101"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.gz", "*.json", "*.yaml"]

[tool.setuptools.packages.find]
exclude = ["*.pyc", "__pycache__"]
namespaces = true
where = ["src"]

[tool.setuptools_scm]
